Remove caching idea and move Type classes from Settings folder.
