+++ CRM/CiviTournament/BillingOrganization.php
@@ -1,5 +1,5 @@
 <?php
-require_once("Settings.php");
+require_once("BillingOrganizationRelationshipType.php");
 
 /**
  * Billing Organization
